Add audit history

+++ SafeguardSpring/src/main/java/com/dqtri/mango/safeguard/controller/CleanupController.java
@@ -1,15 +1,19 @@
 package com.dqtri.mango.safeguard.controller;
 
+import com.dqtri.mango.safeguard.annotation.AuthenticationApiResponses;
+import com.dqtri.mango.safeguard.audit.AuditAction;
+import com.dqtri.mango.safeguard.exception.LockedException;
 import com.dqtri.mango.safeguard.model.LoginAttempt;
 import com.dqtri.mango.safeguard.model.SafeguardUser;
 import com.dqtri.mango.safeguard.model.Submission;
 import com.dqtri.mango.safeguard.repository.LoginAttemptRepository;
 import com.dqtri.mango.safeguard.repository.SubmissionRepository;
 import com.dqtri.mango.safeguard.repository.UserRepository;
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.responses.ApiResponse;
+import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.tags.Tag;
-import jakarta.persistence.EntityNotFoundException;
 import jakarta.transaction.Transactional;
-import jakarta.validation.constraints.NotNull;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Profile;
 import org.springframework.http.ResponseEntity;
@@ -36,13 +40,22 @@ public class CleanupController {
     @DeleteMapping
     public ResponseEntity<String> cleanUpData() {
         // Logic to clean up data
-        return ResponseEntity.ok("Data cleaned up successfully");
+        return ResponseEntity.noContent().build();
     }
 
+    @Operation(summary = "Cleanup testing user", description = "Delete a testing user.")
+    @AuthenticationApiResponses
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "204", description = "Successful cleanup a testing user"),
+    })
+    @Transactional
+    @AuditAction("CLEANUP_USER")
     @DeleteMapping("/users/{email}")
     @PreAuthorize("hasRole('ADMIN')")
-    @Transactional
     public ResponseEntity<Void> cleanupUsers(@PathVariable String email) {
+        if ("admin1@dqtri.com".equalsIgnoreCase(email)){
+            throw new LockedException("Cannot remove this user");
+        }
         List<Submission> submissions = submissionRepository.findAllBySubmitterEmail(email);
         submissionRepository.deleteAll(submissions);
 
@@ -52,21 +65,35 @@ public class CleanupController {
         return ResponseEntity.noContent().build();
     }
 
+    @Operation(summary = "Cleanup testing login attempt", description = "Delete a testing login attempt.")
+    @AuthenticationApiResponses
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "204", description = "Successful cleanup a testing login attempt"),
+    })
+    @Transactional
+    @AuditAction("CLEANUP_LOGIN-ATTEMPT")
     @DeleteMapping("/login-attempt/{email}")
     @PreAuthorize("hasRole('ADMIN')")
-    @Transactional
     public ResponseEntity<Void> cleanupLoginAttempt(@PathVariable String email) {
         Optional<LoginAttempt> byEmail = loginAttemptRepository.findByEmail(email);
         byEmail.ifPresent(loginAttemptRepository::delete);
+
         return ResponseEntity.noContent().build();
     }
 
+    @Operation(summary = "Cleanup testing submission", description = "Delete a testing submission.")
+    @AuthenticationApiResponses
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "204", description = "Successful cleanup a testing submission"),
+    })
+    @Transactional
+    @AuditAction("CLEANUP_SUBMISSION")
     @DeleteMapping("/submissions/{id}")
     @PreAuthorize("hasRole('ADMIN')")
-    @Transactional
-    public ResponseEntity<Void> cleanupLoginAttempt(@PathVariable Long id) {
+    public ResponseEntity<Void> cleanupSubmissions(@PathVariable Long id) {
         Optional<Submission> byId = submissionRepository.findById(id);
         byId.ifPresent(submissionRepository::delete);
+
         return ResponseEntity.noContent().build();
     }
 }
